% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodetector.data.r
\name{geodetector.data}
\alias{geodetector.data}
\title{Prepare data for Geodetector}
\usage{
geodetector.data(object, ssh.lyr, ssh.id, ssh.name=NULL)
}
\arguments{
\item{object}{A POINT \code{sf} object to be joint. Its geometry should be consistent with the sampling layer, but it may have additional attributes from the SSH layer(s).}

\item{ssh.lyr}{A POLYGON \code{sf} object used as the candidate SSH layer. Its strata ID will be linked to \code{object}.}

\item{ssh.id}{A \code{string} denoting the attribute name of strata ID in the SSH layer.}

\item{ssh.name}{A \code{string} denoting the new attribute name assigned to the strata ID in the output object. If \code{NULL}, the attribute name will be the same as that in the old SSH layer.}
}
\description{
\code{geodetector.data} joins a candidate SSH layer to an existing layer.
}
\seealso{
\link[https://cran.r-project.org/package=geodetector]{geodetector R package}
}
