% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.shp.r
\name{load.shp}
\alias{load.shp}
\title{Load shapefile into \code{sf}}
\usage{
load.shp(path, file)
}
\arguments{
\item{path}{Directory of the shapefile.}

\item{file}{Name of the shapefile, minus extension (.shp).}
}
\description{
\code{load.shp} converts a shapefile into a \code{sf} object.
}
