% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ci.r
\name{plot.ci}
\alias{plot.ci}
\title{Plot confidence interval of the Sandwich estimates}
\usage{
plot.ci(object)
}
\arguments{
\item{object}{An \code{sf} object generated from the \code{sandwich.confint} function.}
}
\description{
\code{plot.ci} plots the lower and upper bounds of the confidence interval of estimated mean value for each reporting unit.
}
