% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.se.r
\name{plot.se}
\alias{plot.se}
\title{Plot standard error of the Sandwich estimates}
\usage{
plot.se(object)
}
\arguments{
\item{object}{An \code{sf} object generated from the \code{sandwich} function.}
}
\description{
\code{plot.se} plots the standard error of estimated mean value for each reporting unit.
}
\references{
Wang, J. F., Haining, R., Liu, T. J., Li, L. F., & Jiang, C. S. (2013). Sandwich estimation for multi-unit reporting on a stratified heterogeneous surface. \emph{Environment and Planning A}, 45(10), 2515-2534. doi:\link[https://doi.org/10.1068/a44710]{10.1068/a44710}
}
