% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.confint.r
\name{sandwich.confint}
\alias{sandwich.confint}
\title{Calculate confidence interval of the Sandwich estimates}
\usage{
sandwich.confint(object, level=.95)
}
\arguments{
\item{object}{An \code{sf} object generated from the \code{sandwich.model} function.}

\item{level}{The confidence level required.}
}
\description{
\code{sandwich.confint} calculates the confidence interval of estimated mean value for each reporting unit.
}
