% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.cv.r
\name{sandwich.cv}
\alias{sandwich.cv}
\title{Perform \emph{k}-fold cross validation}
\usage{
sandwich.cv(sampling.lyr, ssh.lyr, reporting.lyr, sampling.attr, k=10)
}
\arguments{
\item{sampling.lyr}{A POINT \code{sf} object used as the sampling layer. It should contain an attribute that is to be interpolated.}

\item{ssh.lyr}{A POLYGON \code{sf} object used as the SSH layer.}

\item{reporting.lyr}{A POLYGON \code{sf} object used as the reporting layer.}

\item{sampling.attr}{A \code{string} denoting the name of the attribute in the sampling layer to be interpolated.}

\item{k}{The number of folds (\code{k} > 1). By default, \code{k} = 10.}
}
\description{
\code{sandwich.cv} perfoms \emph{k}-fold cross validation to evaluate the overall model accuracy and outputs the average root mean square error (RMSE).
}
