% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.model.r
\name{sandwich.model}
\alias{sandwich.model}
\title{Perform Sandwich model-based mapping}
\usage{
sandwich(sampling.lyr, ssh.lyr, reporting.lyr, sampling.attr)
}
\arguments{
\item{sampling.lyr}{A POINT \code{sf} object used as the sampling layer. It should contain an attribute that is to be interpolated.}

\item{ssh.lyr}{A POLYGON \code{sf} object used as the SSH layer.}

\item{reporting.lyr}{A POLYGON \code{sf} object used as the reporting layer.}

\item{sampling.attr}{A \code{string} denoting the name of the attribute in the sampling layer to be interpolated.}
}
\description{
\code{sandwich.model} estimate the mean value of an attribute and its standard error for each reporting unit using the Sandwich method documented by Wang et al. (2013).
}
\references{
Wang, J. F., Haining, R., Liu, T. J., Li, L. F., & Jiang, C. S. (2013). Sandwich estimation for multi-unit reporting on a stratified heterogeneous surface. \emph{Environment and Planning A}, 45(10), 2515-2534. doi:\link[https://doi.org/10.1068/a44710]{10.1068/a44710}
}
