% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.data.shp.r, R/ssh.data.txt.r
\name{ssh.data.shp}
\alias{ssh.data.shp}
\alias{ssh.data.txt}
\title{Prepare shapefile data for Geodetector}
\usage{
ssh.data.shp(object, ssh.lyr, ssh.id, ssh.name=NULL)

ssh.data.txt(object)
}
\arguments{
\item{object}{A list of 2 data frames generated by \code{load.data.txt}, including a file linking sampling and SSH layers linking reporting and SSH layers and a file.}

\item{ssh.lyr}{A POLYGON \code{sf} object used as the candidate SSH layer. Its strata ID will be linked to \code{object}.}

\item{ssh.id}{A \code{string} denoting the attribute name of strata ID in the SSH layer.}

\item{ssh.name}{A \code{string} denoting the new attribute name assigned to the strata ID in the output object. If \code{NULL}, the attribute name will be the same as that in the old SSH layer.}
}
\description{
\code{ssh.data.shp} joins a candidate SSH layer to an existing layer.

\code{ssh.data.txt} joins a candidate SSH layer to an existing layer.
}
\seealso{
\link[https://cran.r-project.org/package=geodetector]{geodetector R package}

\link[https://cran.r-project.org/package=geodetector]{geodetector R package}
}
