% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc.r
\docType{data}
\name{bc}
\alias{bc}
\title{Breast cancer incidence in mainland China}
\description{
The \code{bc} dataset consists of three \code{sf} objects:
\itemize{
\item\code{bc.sampling}: Breast cancer incidence at 271 sampling units in mainland China, which serves as the sampling layer.
\item\code{bc.ssh}: A classification map with 20 strata that serves as a candidate SSH layer.
\item\code{bc.reporting}: The administrative division of 32 provices in mainland China that served as the reporting layer.}
}
\examples{
library(sf)
library(ggplot2)
data("bc.sampling")
summary(bc.sampling$Incidence_)
ggplot(data=bc.sampling) + geom_sf(aes(color=Incidence_))
}
