% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.cv.r
\name{sandwich.cv}
\alias{sandwich.cv}
\title{Perform \emph{k}-fold cross validation}
\usage{
sandwich.cv(object, sampling.attr, k=10, type="shp", ssh.id.col=NULL, ssh.weights=NULL)
}
\arguments{
\item{object}{When \code{type="shp"}, \code{object} is a list of 3 \code{sf} objects generated by \code{load.data.shp}, including a POINT \code{sf} object used as the sampling layer, a POLYGON \code{sf} object used as the SSH layer, and a POLYGON \code{sf} object used as the SSH layer. When \code{type="txt"}, \code{object} is a list of 2 data frames generated by \code{load.data.txt}, including a file linking sampling and SSH layers linking reporting and SSH layers and a file.}

\item{sampling.attr}{A \code{string} denoting the name of the attribute in the sampling layer to be interpolated.}

\item{k}{The number of folds (\code{k} > 1). By default, \code{k} = 10.}

\item{type}{A \code{string} denoting the type of input data. \code{type="shp"} denotes shapefiles, and \code{type="txt"} denotes text files. By default, \code{type="shp"}.}

\item{ssh.id.col}{A \code{string} denoting the column that specifies which stratum each sampling unit falls into in the file linking sampling and SSH layers.}

\item{reporting.id.col}{A \code{string} denoting the column that specifies which reporting unit each sampling unit falls into in the file linking sampling and SSH layers.}

\item{ssh.weights}{A \code{list} that specifies the strata in the SSH layer and and their corresponding columns of weights in the file linking reporting and SSH layers.}
}
\description{
\code{sandwich.cv} perfoms \emph{k}-fold cross validation to evaluate the overall model accuracy and outputs the average root mean square error (RMSE).
}
